# mxsuite-api/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: MXSuite API
  description: |
    MXSuite Data Onboarding Platform — REST API.

    ## Authentication
    All endpoints (except `/api/auth/login` and `/actuator/health`) require a Bearer JWT token
    in the `Authorization` header.

    Platform users (`PLATFORM_ADMIN`, `PLATFORM_SUPPORT`) may include an `X-Tenant-Id` header
    to operate within a specific customer tenant's context.

    ## Pagination
    List endpoints accept Spring Data `Pageable` query parameters:
    - `page` — zero-based page index (default `0`)
    - `size` — page size (default `20`)
    - `sort` — property,direction (e.g. `createdAt,desc`)

    ## Multi-Tenancy
    All tenant-scoped resources are automatically filtered by the tenant context
    derived from the authenticated user's JWT token.
  version: 0.1.0
  contact:
    name: MXSuite Platform Team
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/audit:
    get:
      tags: [Audit]
      summary: List audit events for current tenant
      parameters:
        - name: platformOnly
          in: query
          required: false
          schema:
            type: boolean
          description: Filter to only platform (GrowthZone staff) actions
          example: true
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Paginated audit event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAuditEvent'
              example:
                content:
                  - id: ae111111-1111-1111-1111-111111111111
                    tenantId: "22222222-2222-2222-2222-222222222222"
                    actorId: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                    actorName: Platform Admin
                    actorRole: PLATFORM_ADMIN
                    platformAction: true
                    action: CREATE
                    entityType: Project
                    entityId: p1111111-1111-1111-1111-111111111111
                    entityName: Member Import 2026
                    beforeState: null
                    afterState:
                      name: Member Import 2026
                      description: Annual member data import from legacy CRM
                    metadata: null
                    timestamp: "2026-07-10T09:00:00Z"
                    ipAddress: "192.168.1.100"
                    traceId: "4bf92f3577b34da6a3ce929d0e0e4736"
                  - id: ae222222-2222-2222-2222-222222222222
                    tenantId: "22222222-2222-2222-2222-222222222222"
                    actorId: b2c3d4e5-f6a7-8901-bcde-f12345678901
                    actorName: Denver Admin
                    actorRole: TENANT_ADMIN
                    platformAction: false
                    action: EXECUTE_DRY_RUN
                    entityType: PlanRun
                    entityId: r1111111-1111-1111-1111-111111111111
                    entityName: Contact Import Mapping
                    beforeState: null
                    afterState: null
                    metadata: null
                    timestamp: "2026-07-15T16:30:00Z"
                    ipAddress: "10.0.0.52"
                    traceId: "7a8b9c0d1e2f3a4b5c6d7e8f9a0b1c2d"
                totalElements: 2
                totalPages: 1
                size: 20
                number: 0
  /api/audit/entity/{entityType}/{entityId}:
    get:
      tags: [Audit]
      summary: Get audit events for a specific entity
      description: Returns the complete audit history for a given entity (e.g., all actions on a specific Plan).
      parameters:
        - name: entityType
          in: path
          required: true
          schema:
            type: string
          example: Plan
        - name: entityId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: pl111111-1111-1111-1111-111111111111
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Audit trail for the entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAuditEvent'
              example:
                content:
                  - id: ae333333-3333-3333-3333-333333333333
                    actorName: Denver Admin
                    actorRole: TENANT_ADMIN
                    platformAction: false
                    action: CREATE
                    entityType: Plan
                    entityId: pl111111-1111-1111-1111-111111111111
                    entityName: Contact Import Mapping
                    timestamp: "2026-07-11T10:00:00Z"
                    traceId: "1a2b3c4d5e6f7a8b9c0d1e2f3a4b5c6d"
                  - id: ae444444-4444-4444-4444-444444444444
                    actorName: Denver Admin
                    actorRole: TENANT_ADMIN
                    platformAction: false
                    action: UPDATE_DEFINITION
                    entityType: Plan
                    entityId: pl111111-1111-1111-1111-111111111111
                    entityName: Contact Import Mapping
                    timestamp: "2026-07-14T14:00:00Z"
                    traceId: "2b3c4d5e6f7a8b9c0d1e2f3a4b5c6d7e"
                  - id: ae555555-5555-5555-5555-555555555555
                    actorName: Platform Admin
                    actorRole: PLATFORM_ADMIN
                    platformAction: true
                    action: PUBLISH
                    entityType: Plan
                    entityId: pl111111-1111-1111-1111-111111111111
                    entityName: Contact Import Mapping
                    timestamp: "2026-07-15T16:00:00Z"
                    traceId: "3c4d5e6f7a8b9c0d1e2f3a4b5c6d7e8f"
                totalElements: 3
                totalPages: 1
                size: 20
                number: 0

  # ──────────────────────────────────────────────
  # Profile (Self-Service)
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /api/auth/login
  parameters:
...
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Zero-based page index
    size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Page size
    sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
      description: "Sort property,direction (e.g. createdAt,desc)"
      example: createdAt,desc
  schemas:
...
    AuditEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        actorId:
          type: string
          format: uuid
        actorName:
          type: string
        actorRole:
          type: string
        platformAction:
          type: boolean
          description: True if action was performed by a platform (GrowthZone) user
        action:
          type: string
          description: "Action type (CREATE, UPDATE, DELETE, PUBLISH, ARCHIVE, EXECUTE_DRY_RUN, EXECUTE_FULL_RUN, UPLOAD, SHARE)"
        entityType:
          type: string
          description: "Entity class name (Tenant, User, Project, Plan, PlanRun, Workspace, ProjectAsset)"
        entityId:
          type: string
          format: uuid
        entityName:
          type: string
          nullable: true
        beforeState:
          type: object
          nullable: true
          description: JSONB snapshot of entity state before change
        afterState:
          type: object
          nullable: true
          description: JSONB snapshot of entity state after change
        metadata:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
        ipAddress:
          type: string
          nullable: true
        traceId:
          type: string
          nullable: true
          description: OpenTelemetry trace ID for distributed trace correlation

    # ── Enums ──
...
    PageAuditEvent:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
...